+++ demisto_sdk/__main__.py
@@ -797,6 +797,10 @@ def run_playbook(**kwargs):
 @click.option(
     "--interactive", help="If passed, then for each output field will ask user interactively to enter the "
                           "description. By default is interactive mode is disabled", is_flag=True)
+@click.option(
+    "-d", "--description_json",
+    help="If passed, user is prompted to input a JSON mapping field names to their descriptions. False by default.",
+    is_flag=True)
 def json_to_outputs_command(**kwargs):
     """Demisto integrations/scripts have a YAML file that defines them.
     Creating the YAML file is a tedious and error-prone task of manually copying outputs from the API result to the