+++ CHANGELOG.md
@@ -6,7 +6,10 @@
 * Added new validation for the `xsoar_config.json` file in the **validate** command.
 * Added a version differences section to readme in **generate-docs** command.
 * Added the *--docs-format* flag in the **integration-diff** command to get the output in README format.
-* Added the *--input-old-version* and *--skip-breaking-changes* flags in the **generate-docs** command to get the details for the breaking section and to skip the breaking changes section.
+* Added the *--input-old-version* and *--skip-breaking-changes* flags in the **generate-docs** command to get the
+  details for the breaking section and to skip the breaking changes section.
+* Added option to enter a dictionary or json of format `[{field_name:description}]` in the **json-to-outputs** command,
+  with the `-d` flag.
 
 # 1.4.0
 * Enable passing a comma-separated list of paths for the `--input` option of the **lint** command.